requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
constantly==15.1.0
crayons==0.4.0
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
h2==3.2.0
hiredis==2.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
ImageHash==4.2.0
incremental==21.3.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itemadapter==0.2.0
itemloaders==1.0.4
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parsel==1.6.0
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
priority==1.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
Protego==0.1.16
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.9.0
pymongo==3.11.4
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
queuelib==1.6.1
redis==3.5.3
requests==2.25.1
scipy==1.6.3
Scrapy==2.5.0
Scrapy-Cookies==0.3
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
statsmodels==0.12.2
terminado==0.10.0
testpath==0.5.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
Twisted==21.2.0
ujson==4.0.2
urllib3==1.26.4
w3lib==1.22.0
wcwidth==0.2.5
webdriver-manager==3.4.1
webdrivermanager==0.10.0
webencodings==0.5.1
widgetsnbextension==3.5.1
zope.interface==5.4.0
